# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BarberHub API",
        "description": "API for BarberHub barbershop booking system",
        "contact": {
            "email": "support@barberhub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
...
        "/owner/dashboard": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "Owner dashboard",
                "description": "Returns barbershop owner dashboard with aggregated metrics for the authenticated user`s barbershop.",
                "operationId": "ownerDashboard",
                "responses": {
                    "200": {
                        "description": "Dashboard data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "barbershop": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "rating": {
                                                            "type": "number",
                                                            "format": "float"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "today_bookings": {
                                                    "properties": {
                                                        "count": {
                                                            "type": "integer",
                                                            "example": 7
                                                        },
                                                        "delta_vs_yesterday": {
                                                            "type": "integer",
                                                            "example": 3
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "week_revenue": {
                                                    "properties": {
                                                        "amount": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 85000
                                                        },
                                                        "delta_pct_vs_prev": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 12.5,
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "new_clients_this_week": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "revenue_last_7_days": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "date": {
                                                                "type": "string",
                                                                "format": "date"
                                                            },
                                                            "revenue": {
                                                                "type": "number",
                                                                "format": "float"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "nearest_bookings": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "client_name": {
                                                                "type": "string"
                                                            },
                                                            "barber_name": {
                                                                "type": "string"
                                                            },
                                                            "scheduled_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "services_count": {
                                                                "type": "integer"
                                                            },
                                                            "total_price": {
                                                                "type": "number",
                                                                "format": "float"
                                                            },
                                                            "status": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "User is not a barbershop owner"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/owner/calendar": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "Owner calendar",
                "description": "Returns bookings grouped by date for a given range (max 31 days).",
                "operationId": "ownerCalendar",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-04-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-04-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendar data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "from": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "to": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "days": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "date": {
                                                                "type": "string",
                                                                "format": "date"
                                                            },
                                                            "count": {
                                                                "type": "integer"
                                                            },
                                                            "bookings": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer"
                                                                        },
                                                                        "client_name": {
                                                                            "type": "string"
                                                                        },
                                                                        "barber_name": {
                                                                            "type": "string"
                                                                        },
                                                                        "scheduled_at": {
                                                                            "type": "string",
                                                                            "format": "date-time"
                                                                        },
                                                                        "total_price": {
                                                                            "type": "number",
                                                                            "format": "float"
                                                                        },
                                                                        "status": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not an owner"
                    },
                    "422": {
                        "description": "Validation error (invalid range or > 31 days)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/owner/bookings": {
            "get": {
                "tags": [
                    "Owner"
                ],
                "summary": "List owner bookings",
                "description": "Paginated list of bookings for the owner`s barbershop with optional status filter",
                "operationId": "ownerBookings",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "pending",
                                "confirmed",
                                "cancelled",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings list"
                    },
                    "403": {
                        "description": "Not an owner"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}